+++ src/main/java/domain/result/RoundBetInfo.java
@@ -1,10 +1,11 @@
 package domain.result;
 
 import domain.player.User;
+import java.math.BigDecimal;
 
-public record RoundBetInfo(int round, User user, double betAmount) {
+public record RoundBetInfo(int round, User user, BigDecimal betAmount) {
 
-    public UserProfit toUserProfit(GameResult gameResult, double profit) {
-        return new UserProfit(this, gameResult,profit);
+    public UserProfit toUserProfit(GameResult gameResult, BigDecimal profit) {
+        return new UserProfit(this, gameResult, profit);
     }
 }
