+++ src/main/java/view/InputParser.java
@@ -1,11 +1,11 @@
 package view;
 
+import domain.result.RoundBetInfo;
+import domain.player.User;
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.IntStream;
-
-import domain.result.RoundBetInfo;
-import domain.player.User;
 
 public class InputParser {
 
@@ -17,7 +17,7 @@ public class InputParser {
 
     public static List<RoundBetInfo> parseToRoundBetInfos(List<String> names, List<Integer> betAmounts, int round) {
         return IntStream.range(0, names.size())
-                .mapToObj(i -> new RoundBetInfo(round, User.from(names.get(i)), betAmounts.get(i)))
+                .mapToObj(i -> new RoundBetInfo(round, User.from(names.get(i)), BigDecimal.valueOf(betAmounts.get(i))))
                 .toList();
     }
 }
